cpukit libfs: Address -Wsign-compare warnings
Also includes a related libio change.

Most of these warnings were between int or ssize_t and size_t.
In particular, various POSIX system calls like read() and write()
return ssize_t and comparing that value with the sizeof a buffer
is a common source of these warnings. Another common source is
using an int as the iterator in a for loop with the limit being
a size_t.

With the type change, some printf() specifiers needed to change also.

+++ cpukit/include/rtems/confdefs/libio.h
@@ -199,7 +199,8 @@ extern "C" {
   #error "CONFIGURE_IMFS_MEMFILE_BYTES_PER_BLOCK must be a power of two between 16 and 1024"
 #endif
 
-const int imfs_memfile_bytes_per_block = CONFIGURE_IMFS_MEMFILE_BYTES_PER_BLOCK;
+const size_t imfs_memfile_bytes_per_block =
+    CONFIGURE_IMFS_MEMFILE_BYTES_PER_BLOCK;
 
 /*
  * Using the default ops if user doesn't configure ops
